extract: logic from RedhatRepositoriesStep
All the logic has been extracted into its context store.
The component has been divided into constituent subcomponents -
heading, table controls, table, empty table, table loading skeleton.

The store has been placed above the template base modal.
