+++ src/features/createAndEditTemplate/redhatRepositories/ui/components/EmptyTable.tsx
@@ -1,0 +1,17 @@
+import { Bullseye } from '@patternfly/react-core';
+import EmptyTableState from 'components/EmptyTableState/EmptyTableState';
+import { useRedhatRepositoriesApi } from '../../store/RedhatRepositoriesStore';
+
+export const EmptyTable = () => {
+  const { searchQuery, setSearchQuery } = useRedhatRepositoriesApi();
+  return (
+    <Bullseye data-ouia-component-id='redhat_repositories_table'>
+      <EmptyTableState
+        notFiltered={searchQuery === ''}
+        clearFilters={() => setSearchQuery('')}
+        itemName='Red Hat repositories'
+        notFilteredBody='No Red Hat repositories match the version and arch'
+      />
+    </Bullseye>
+  );
+};
